+++ pyincore/globals.py
@@ -9,15 +9,15 @@ from logging import config as logging_config
 import logging
 import os
 
-PACKAGE_VERSION = "0.5.4"
+PACKAGE_VERSION = "0.5.5"
 
 INCORE_API_PROD_URL = "https://incore.ncsa.illinois.edu"
+TEST_INCORE_API_PROD_URL = "http://incore.ncsa.illinois.edu:31888"
 INCORE_API_DEV_URL = "https://incore-dev-kube.ncsa.illinois.edu"
-
-KONG_INCORE_API_PROD_URL = "https://incore2-services.ncsa.illinois.edu"
-KONG_INCORE_API_DEV_URL = "https://incore2-services-dev.ncsa.illinois.edu"
+TEST_INCORE_API_DEV_URL = "http://incore-dev-kube.ncsa.illinois.edu:31888"
 
 KEYCLOAK_AUTH_PATH = "/auth/realms/In-core/protocol/openid-connect/token"
+KEYCLOAK_USERINFO_PATH = "/auth/realms/In-core/protocol/openid-connect/userinfo"
 CLIENT_ID = "react-auth"
 INCORE_LDAP_TEST_USER_INFO = "{\"preferred_username\": \"incrtest\"}"